% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{add_stopping_prob}
\alias{add_stopping_prob}
\title{Calculate stopping probabilities}
\usage{
add_stopping_prob(x)
}
\arguments{
\item{x}{Object of class gsnb}
}
\value{
Object of class gsnb
}
\description{
Calculate analyses specific stopping probabilities and
expected information level
}
\keyword{internal}
