% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gsnb.R
\name{check_gsnb}
\alias{check_gsnb}
\title{Check object of class gsnb}
\usage{
check_gsnb(x)
}
\arguments{
\item{x}{object of class gsnb}
}
\description{
Performs validity checks of object of class gsnb
}
\keyword{internal}
