% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{design_gsnb}
\alias{design_gsnb}
\title{Design a group sequential trial with negative binomial data}
\usage{
design_gsnb(rate1, rate2, shape, power_gs, timing, esf = esf_obrien,
  ratio_H0 = 1, sig_level, random_ratio = 1, t_recruit1 = NULL,
  t_recruit2 = NULL, study_period = NULL, accrual_period = NULL,
  followup_max = NULL)
}
\arguments{
\item{rate1}{numeric; assumed rate of treatment group 1 in the alternative}

\item{rate2}{numeric; assumed rate of treatment group 2 in the alternative}

\item{shape}{numeric; shape parameter}

\item{power_gs}{numeric; target power of group sequential design}

\item{timing}{numeric vector; 0 < \code{timing[1]} < ... < \code{timing[K]} = 1
with \code{K} the number of analyses, i.e. (K-1) interim analyses and final analysis}

\item{esf}{function; error spending function}

\item{ratio_H0}{numeric; postive number denoting the rate ratio rate_1/rate_2
under the null hypothesis, i.e. the non-inferiority or superiority margin}

\item{sig_level}{numeric; Type I error / significance level}

\item{random_ratio}{numeric; randomization ratio n1/n2}

\item{t_recruit1}{numeric vector; recruit (i.e. study entry) times in group 1}

\item{t_recruit2}{numeric vector; recruit (i.e. study entry) times in group 2}

\item{study_period}{numeric; study duration}

\item{accrual_period}{numeric; accrual period}

\item{followup_max}{numeric; maximum exposure time of a patient}
}
\value{
A list with class "GSNBdesign" containing the following components:
\item{rate1}{as input}
\item{rate2}{as input}
\item{shape}{as input}
\item{power_gs}{as input}
\item{timing}{as input}
\item{ratio_H0}{as input}
\item{ratio_H1}{ratio \code{rate1}/\code{rate2}}
\item{sig_level}{as input}
\item{random_ratio}{as input}
\item{t_recruit1}{as input}
\item{t_recruit2}{as input}
\item{study_period}{as input}
\item{followup_max}{as input}
\item{max_info}{ maximum information}
\item{power_fix}{power of fixed design}
\item{critical}{critical values for each look}
\item{reject_prob}{data frame with rejection (i.e. boundary crossing) probabilities}
}
\description{
Design a group sequential trial with negative binomial data
}
\examples{
# Calculate the sample sizes for a given accrual period and study period
out <- design_gsnb(rate1 = 0.0875, rate2 = 0.125, shape = 5, power_gs = 0.8,
                   timing = c(0.5, 1), esf = esf_obrien,
                   ratio_H0 = 1, sig_level = 0.025,
                   study_period = 3.5, accrual_period = 1.25, random_ratio = 1)
out

# Calculate study period for given recruitment times
expose <- seq(0, 1.25, length.out = 1042)
out <- design_gsnb(rate1 = 0.0875, rate2 = 0.125, shape = 5, power_gs = 0.8,
                   timing = c(0.5, 1), esf = esf_obrien,
                   ratio_H0 = 1, sig_level = 0.025,
                   t_recruit1 = expose, t_recruit2 = expose, random_ratio = 1)
out

# Calculate sample size for a fixed exposure time
out <- design_gsnb(rate1 = 0.0875, rate2 = 0.125, shape = 5, power_gs = 0.8,
                   timing = c(0.5, 1), esf = esf_obrien,
                   ratio_H0 = 1, sig_level = 0.025,
                   followup_max = 0.5, random_ratio = 1)
}
