% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spending_functions.R
\name{esf_obrien}
\alias{esf_obrien}
\title{esf_obrien}
\usage{
esf_obrien(t, sig_level, ...)
}
\arguments{
\item{t}{numeric; Non-negative information ratio}

\item{sig_level}{numeric; significance level}

\item{...}{optional arguments}
}
\value{
numeric
}
\description{
Error spending function mimicking O'Brien & Fleming critical values
}
