% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spending_functions.R
\name{esf_pocock}
\alias{esf_pocock}
\title{esf_pocock}
\usage{
esf_pocock(t, sig_level, ...)
}
\arguments{
\item{t}{numeric; Non-negative information ratio}

\item{sig_level}{numeric; significance level}

\item{...}{optional arguments}
}
\value{
numeric
}
\description{
Error spending function mimicking Pococks critical values
}
