% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{get_covar}
\alias{get_covar}
\title{get_covar}
\usage{
get_covar(timing)
}
\arguments{
\item{timing}{numeric vector; 0 < \code{timing[1]} < ... < \code{timing[k]}}
}
\description{
Get canonical form covariance matrix
}
\keyword{internal}
