% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical_values.R
\name{get_critical_values}
\alias{get_critical_values}
\title{Calculate critical values}
\usage{
get_critical_values(timing, alloc_error, prev_critical = NULL)
}
\arguments{
\item{timing}{numeric vector; 0 < \code{timing[1]} < ... < \code{timing[k]} with
\code{k} the index and \code{timing[k]} the information
ratio of the current analysis}

\item{alloc_error}{Numeric vector with allocated type I error.}

\item{prev_critical}{Critical values of previous analyses}
}
\value{
numeric; critical values
}
\description{
Calculate either critical value for analysis at stage k based on
given previous values or calculate set of critical values for multiple analysis.
}
\details{
If \code{alloc_error} has length 1 and \code{prev_critical} is \code{NULL},
the corresponding normal quantile, i.e. the critical value for the first interim,
is returned.
If length of \code{alloc_error} and \code{timing} is identical and \code{prev_critical}
is \code{NULL}, the vector of critical values for the analyses at times \code{timing}
is returned.
If \code{alloc_error} has length 1 and \code{timing} has one entry more than
\code{prev_critical}, the critical value for the analysis at time \code{timing[k]}
is returned.
}
