% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{get_rejectprob_gsnb}
\alias{get_rejectprob_gsnb}
\title{Calcualte rejection probabilities}
\usage{
get_rejectprob_gsnb(rate_ratio, ratio_H0 = 1, critical, max_info, timing,
  n_vec = NULL, study_period_vec = NULL)
}
\arguments{
\item{rate_ratio}{The rate ratio for which the rejection probability is calculated.}

\item{ratio_H0}{The rate ratio in the null hypothesis against which the test is performed.}

\item{critical}{The vector of critical values used for the test. Must have
the same length as argument \code{timing}.}

\item{max_info}{The maximum information of the study.}

\item{timing}{The vector of information times at which the tests are performed.}

\item{n_vec}{An optional vector with \code{n_vec[i]} indicating the number of
subjects recruited at information time \code{timing[i]}.
Must have the same length as argument \code{timing}.}

\item{study_period_vec}{An optional vector with \code{study_period_vec[i]}
indicating the study time at information time \code{timing[i]}.
Must have the same length as argument \code{timing}.}
}
\value{
data frame with rejection (i.e. boundary crossing) probabilities
}
\description{
Calcualte analyses specific rejection probabilities and
expected information level, study duration, and sample size
}
\keyword{internal}
