% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{get_timepoints_gsnb}
\alias{get_timepoints_gsnb}
\title{Time point of looks}
\usage{
get_timepoints_gsnb(rate1, rate2, shape, t_recruit1, t_recruit2, timing,
  followup1, followup2)
}
\arguments{
\item{rate1}{numeric; rate in treatment group 1}

\item{rate2}{numeric; rate in treatment group 2}

\item{shape}{numeric; shape parameter}

\item{t_recruit1}{numeric vector; recruit (i.e. study entry) times in group 1}

\item{t_recruit2}{numeric vector; recruit (i.e. study entry) times in group 2}

\item{timing}{numeric vector with entries in (0,1]; information times of looks}

\item{followup1}{numeric vector; final individual follow-up times in treatment group 1}

\item{followup2}{numeric vector; final individual follow-up times in treatment group 2}
}
\value{
numeric; vector with the time points of the looks
}
\description{
Calculate the time points of looks given the information time
}
