% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{gscounts}
\alias{gscounts}
\alias{gscounts-package}
\title{gscounts}
\description{
Design and monitoring of group sequential designs with negative binomial data.
}
\author{
Tobias Muetze <tobias.muetze@outlook.com>
}
