% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spending_functions.R
\name{obrien}
\alias{obrien}
\title{obrien}
\usage{
obrien(t, sig_level, ...)
}
\arguments{
\item{t}{numeric; Non-negative information ratio}

\item{sig_level}{numeric; significance level}

\item{...}{optional arguments}
}
\value{
numeric
}
\description{
Error spending function mimicking O'Brien & Fleming critical values
}
\examples{
# O'Brien-Fleming-type error spending function
obrien(t = c(0.5, 1), sig_level = 0.025)
}
