% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.GSNBpower}
\alias{plot.GSNBpower}
\title{plot.GSNBpower}
\usage{
\method{plot}{GSNBpower}(x)
}
\arguments{
\item{x}{an object of class \emph{GSNBpower}}
}
\value{
Returns a gg-plot, i.e. an instance with classes \emph{gg} and \emph{ggplot}
}
\description{
Plots an object of class GSNBdesign
}
\details{
The following entries of the objects of class \emph{GSNBdesign}
are considered to calculate the power: \code{ratio_H0}, \code{ratio_H1},
\code{timing}, \code{sig_level}, \code{esf}
}
