% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spending_functions.R
\name{pocock}
\alias{pocock}
\title{pocock}
\usage{
pocock(t, sig_level, ...)
}
\arguments{
\item{t}{numeric; Non-negative information ratio}

\item{sig_level}{numeric; significance level}

\item{...}{optional arguments}
}
\value{
numeric
}
\description{
Error spending function mimicking Pococks critical values
}
\examples{
# Pocock-type error spending function
pocock(t = c(0.5, 1), sig_level = 0.025)
}
