% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power_gsnb}
\alias{power_gsnb}
\title{power_gsnb}
\usage{
power_gsnb(ratio_H1, max_info = NULL, power_gs = NULL, timing,
  esf = esf_obrien, ratio_H0 = 1, sig_level, ...)
}
\arguments{
\item{ratio_H1}{numeric; postive number denoting the rate ratio rate_1/rate_2
under the alternative hypothesis}

\item{max_info}{The maximum information of the study.}

\item{power_gs}{numeric; the target power of the group sequential design}

\item{timing}{numeric vector; 0 < timing[1] < ... < timing[K] = 1 with K the
number of analyses, i.e. (K-1) interim analyses and final analysis}

\item{esf}{An error spending function}

\item{ratio_H0}{numeric; postive number denoting the rate ratio rate_1/rate_2
under the null hypothesis, i.e. the non-inferiority or superiority margin}

\item{sig_level}{numeric; the total significance level}

\item{...}{additional arguments for the error spending function esf}
}
\value{
A list with class "GSNBpower" containing the following components:
\item{ratio_H0}{as input}
\item{ratio_H1}{as input}
\item{timing}{as input}
\item{sig_level}{as input}
\item{max_info}{as input if not NULL; otherweise calculated maximum information}
\item{power_gs}{as input if not NULL; otherweise calculated power}
\item{power_fix}{power of fixed design}
\item{critical}{critical values for each look}
}
\description{
Calculate power of group sequential design with
negative binomial endpoints
}
