% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.gsnb}
\alias{print.gsnb}
\title{print.gsnb}
\usage{
\method{print}{gsnb}(x, ...)
}
\arguments{
\item{x}{an object of class gsnb}

\item{...}{optional arguments to print or plot methods}
}
\description{
print method for instance of class gsnb
}
