% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.GSNBpower}
\alias{print.GSNBpower}
\title{print.GSNBpower}
\usage{
\method{print}{GSNBpower}(x, ...)
}
\arguments{
\item{x}{an object of class GSNBpower}

\item{...}{optional arguments to print or plot methods}
}
\description{
print method for instance of class GSNBpower
}
