% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.nb}
\alias{print.nb}
\title{print.nb}
\usage{
\method{print}{nb}(x, ...)
}
\arguments{
\item{x}{an object of class nb}

\item{...}{optional arguments to print or plot methods}
}
\description{
print method for instance of class nb
}
