% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{samplesize_from_followup}
\alias{samplesize_from_followup}
\title{samplesize_from_followup}
\usage{
samplesize_from_followup(max_info, random_ratio, rate1, rate2, shape,
  followup_max)
}
\description{
Determine the sample size from a given follow-up time
}
\keyword{internal}
