% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{samplesize_from_periods}
\alias{samplesize_from_periods}
\title{samplesize_from_periods}
\usage{
samplesize_from_periods(max_info, accrual_period, study_period,
  random_ratio, rate1, rate2, shape, accrual_speed)
}
\description{
Determine the total sample size for given accrual and study period
}
\keyword{internal}
