% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{studyperiod_from_recruit}
\alias{studyperiod_from_recruit}
\title{studyperiod_from_recruit}
\usage{
studyperiod_from_recruit(max_info, random_ratio, rate1, rate2, shape,
  t_recruit1, t_recruit2)
}
\description{
Determine the study period for given recruitment times
}
\keyword{internal}
